refactor: extract LogoutPage to auth feature directory
