+++ src/features/auth/pages/index.ts
@@ -1,1 +1,2 @@
 export * from "./PendingApprovalPage";
+export * from "./LogoutPage";
